infra/requirements.txt
aws-cdk-lib==2.161.1
pydantic==2.12.0